+++ MyFitApp_main/AuditMicroservice/src/main/java/itacad/aliaksandrkryvapust/auditmicroservice/manager/api/IAuditManager.java
@@ -1,6 +1,6 @@
 package itacad.aliaksandrkryvapust.auditmicroservice.manager.api;
 
-import itacad.aliaksandrkryvapust.auditmicroservice.core.dto.AuditDto;
+import itacad.aliaksandrkryvapust.auditmicroservice.core.dto.input.AuditDto;
 import itacad.aliaksandrkryvapust.auditmicroservice.core.dto.output.AuditDtoOutput;
 
 public interface IAuditManager extends IManager<AuditDto,AuditDtoOutput>{
